neon: extract isSameUrlWithoutQuery util and add test
