+++ apps/donations/neon-crm-api/__tests__/utils.test.ts
@@ -1,6 +1,10 @@
 import { it, expect, describe } from "vitest";
 
-import { fixDateTimezone, fixTimestampsTimezone } from "../utils";
+import {
+  fixDateTimezone,
+  fixTimestampsTimezone,
+  isSameUrlWithoutQuery,
+} from "../utils";
 import type { Timestamps } from "../schema";
 
 const timezone = "America/Chicago";
@@ -37,4 +41,13 @@ describe("fixTimestampsTimezone", () => {
     expect(fixed.createdBy).toBe("Some Name");
     expect(fixed.lastModifiedBy).toBe("Some other Name");
   });
+});
+
+describe("isSameUrlWithoutQuery", () => {
+  const a = "https://example.com/api/v1/donations?uuid=123";
+  const b = "https://example.com/api/v1/donations?uuid=456";
+  const c = "https://example.com/api/v1/donations?uuid=123&other=value";
+
+  expect(isSameUrlWithoutQuery(a, b, ["uuid"])).toBe(true);
+  expect(isSameUrlWithoutQuery(a, c, ["uuid"])).toBe(false);
 });
