+++ apps/donations/neon-crm-api/utils.ts
@@ -21,3 +21,22 @@ export const fixTimestampsTimezone = (
   createdDateTime: fixDateTimezone(timestamps.createdDateTime, zone),
   lastModifiedDateTime: fixDateTimezone(timestamps.lastModifiedDateTime, zone),
 });
+
+const cleanSearchParams = (
+  urlString: string,
+  paramsToExclude: string[] = [],
+) => {
+  const url = new URL(urlString);
+  const params = url.searchParams;
+  paramsToExclude.forEach((param) => params.delete(param));
+  params.sort();
+  return `${url.origin + url.pathname}?${params.toString()}`;
+};
+
+export const isSameUrlWithoutQuery = (
+  a: string,
+  b: string,
+  paramsToExclude: string[] = [],
+) =>
+  cleanSearchParams(a, paramsToExclude) ===
+  cleanSearchParams(b, paramsToExclude);
